# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "BugHunter API",
    "description": "API for network bug and vulnerability scanning."
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/scan/ports": {
      "post": {
        "tags": ["Network Scanner"],
        "summary": "Scan ports on a target host",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "target": {
                  "type": "string",
                  "example": "127.0.0.1"
                },
                "ports": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  },
                  "example": [80, 443, 22]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Port scan results",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/scan/vulnerabilities": {
      "post": {
        "tags": ["Vulnerability Scanner"],
        "summary": "Scan for software vulnerabilities",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "software": {
                  "type": "string",
                  "example": "nginx:1.21"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vulnerability scan results",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/scan/web": {
      "post": {
        "tags": ["Web Scanner"],
        "summary": "Scan a web application for vulnerabilities",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "target": {
                  "type": "string",
                  "example": "http://example.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Web application scan results",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/scan/dns": {
      "post": {
        "tags": ["DNS Scanner"],
        "summary": "Scan a domain for DNS configurations",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "domain": {
                  "type": "string",
                  "example": "example.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "DNS scan results",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    }
  }
}